+++ web/admin/src/views/tenant/detail/common.css
@@ -1,0 +1,16 @@
+.money {
+  font-weight: 600;
+  color: var(--color-success-6);
+}
+.table-footer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+  margin-top: 16px;
+  padding-top: 16px;
+  border-top: 1px solid var(--ta-border-light);
+}
+.table-footer .table-stats {
+  margin-bottom: 0;
+}
